+++ src/full/Agda/Interaction/Options/Warnings.hs
@@ -174,6 +174,7 @@ errorWarnings = Set.fromList
   , ConstructorDoesNotFitInData_
   , OverlappingTokensWarning_
   , PragmaCompiled_
+  , RecursiveRecordNeedsInductivity_
   , SafeFlagPostulate_
   , SafeFlagPragma_
   , SafeFlagNonTerminating_
@@ -395,6 +396,7 @@ data WarningName
   | InvalidDisplayForm_
   | UnusedVariablesInDisplayForm_
   | RewritesNothing_
+  | RecursiveRecordNeedsInductivity_
   | WithClauseProjectionFixityMismatch_
   | WithoutKFlagPrimEraseEquality_
   | ConflictingPragmaOptions_
@@ -642,6 +644,7 @@ warningNameDescription = \case
   InvalidDisplayForm_              -> "Invalid display forms."
   UnusedVariablesInDisplayForm_    -> "Bound but unused variables in display forms."
   TooManyArgumentsToSort_          -> "Extra arguments given to a sort."
+  RecursiveRecordNeedsInductivity_ -> "Recursive records declared neither `inductive' nor `coinductive'."
   RewritesNothing_                 -> "`rewrite' clauses that do not fire."
   WithClauseProjectionFixityMismatch_ -> "With clauses using projections in different fixities than their parent clauses."
   WithoutKFlagPrimEraseEquality_   -> "Uses of `primEraseEquality' with the without-K flags."